1155: DeprecationWarning an integer is required self.progressBar.setValue

+++ mantidimaging/gui/dialogs/async_task/view.py
@@ -48,7 +48,7 @@ class AsyncTaskDialogView(BaseDialogView):
             self.infoText.setText(message)
 
         # Update progress bar
-        self.progressBar.setValue(progress * 1000)
+        self.progressBar.setValue(int(progress * 1000))
 
     def show_delayed(self, timeout):
         self.show_timer.singleShot(timeout, self.show_from_timer)
